+++ src/main/java/com/backend/moamoa/global/bean/swagger/SwaggerConfig.java
@@ -17,17 +17,15 @@ import springfox.documentation.schema.AlternateTypeRules;
 import springfox.documentation.service.ApiInfo;
 import springfox.documentation.spi.DocumentationType;
 import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-
 @Configuration
 @EnableWebMvc
 @RequiredArgsConstructor
-public class SwaggerConfiguration {
+public class SwaggerConfig {
     private final TypeResolver typeResolver;
 
     @Bean
